Fixed shadow issue
Everything finally works! The sphere now casts a shadow on the ground. However, since all the light-beams are seen as perpendicular the shadow on the ground disappears when illuminating the sphere from below. This is different from reality where a light source close to the ground illuminating the sphere from below would still cause the sphere to cast a shadow on the ground on the other side.